refactor(code-intelligence): adopt modular security/ package, shim monolith (#9856)
- security/constants.py: restore local definitions for DEBUG_MODE_VARS,
  PICKLE_MODULES, YAML_LOADER_ARGS, LOAD_FUNCS, VALIDATION_FUNCS,
  VALIDATION_ATTRS, SecuritySeverity, VulnerabilityType, OWASP_MAPPING,
  WEAK_HASH_ALGORITHMS, WEAK_ENCRYPTION; re-export only the three names
  that actually landed in autobot_shared/ssot_constants (PLACEHOLDER_PATTERNS,
  HTTP_METHODS, INSECURE_RANDOM_FUNCS) — fixes ImportError on package import
- security_analyzer.py: convert 1,278-line monolith into a thin re-export
  shim over code_intelligence.security; all 3 live consumers unchanged
- shared/ast_cache.py: fix two zero-capacity bugs — infinite while loop
  in _store_parsed_ast and StopIteration in _read_file_content when
  ast_cache_max_size / content_cache_max_size == 0 (pre-existing #9784)
- code_intelligence/conftest.py: un-stub code_intelligence.security_analyzer
  in the test environment (parent conftest stubs it as MagicMock to avoid
  loading the heavy __init__.py chain; local conftest loads the real package
  so security_analyzer_test.py exercises actual detection logic)

Equivalence: 39 VulnerabilityType members confirmed; all 21 monolith tests
pass through the shim (security_analyzer_test.py unchanged).
